Move add_story_and_content_download() to crawler-fetcher
